Package structure breaking climate fixes
