+++ src/main/java/ru/mail/polis/lsm/artem_drozdov/util/File.java
@@ -1,0 +1,25 @@
+package ru.mail.polis.lsm.artem_drozdov.util;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+
+final public class File {
+    public static void rename(Path file, Path tmpFile) throws IOException {
+        Files.deleteIfExists(file);
+        Files.move(tmpFile, file, StandardCopyOption.ATOMIC_MOVE);
+    }
+
+    private static Path resolveWithExt(Path file, String ext) {
+        return file.resolveSibling(file.getFileName() + ext);
+    }
+
+    public static Path getIndexFile(Path file) {
+        return resolveWithExt(file, ".idx");
+    }
+
+    public static Path getTmpFile(Path file) {
+        return resolveWithExt(file, ".tmp");
+    }
+}
